Refactoring log utilities into logUtil.go.
